Refactor templater to accept modules better
I have been struggling in trying to make it a really intuitive interface
in making the structures for pages, etc. I don't love the `_modules`
layout but it's not a big deal. Something that will be easy to manage in
the future if I end up hating it.

I also changed the name of `fill_template` to just `fill` because it's
obvious.